requirements.txt
numpy~=1.25.2
uproot~=5.0.11
pandas~=2.1.0
matplotlib~=3.7.2
scipy~=1.11.2
lmfit~=1.2.2
torch~=2.0.1
tqdm~=4.66.1
pyro-ppl~=1.8.6
setuptools